// src/interfaces/archivoconfirmacion-prima.model.ts
export class ArchivoConfirmacionPrima{
    NumPoliza: number;
    MtoPrima: number;
    FechaInicio?: string;
    FechaFin?: string;
    Usuario: string;
    Fecha: string;
    Hora: string;
    NombreArchivo: string;
    variableUsuario: string;
}

export class ArchivoConfirmacionPrimaXML{
    
    NumOperacion: number;    
    CUSPP: string;
    NumPoliza: string;
    MtoSumPension: number;
    MtoRentaEsc: number;
    MtoPensionRT: number;
    MtoPension: number;
    MtoPriRecAFP: number;
    MtoPriRec: number;
    IniVigencia: string;
    MesDif: number;
    MesEsc: number;
    TipPension: string;
}